geojsontojson: Add meta output, with output field coverage
Will be adding more to meta later, like errors on conversion.

Added field coverage because thought would be helpful for
https://github.com/Open-Telecoms-Data/cove-ofds/issues/6
but actually I added to wrong conversion!
Leaving as may be useful and that is simpler than removing.

Also fix: bug in get_json() that meant it could not be called twice